fix : fixes the parsing of json to throw exception.
The json received in case of any error is of the type `{error: {code: -32602, message: This payment is destined for ourselves. Self-payments are not supported}, id: (null)commando:pay#106, jsonrpc: 2.0}`.
For now the exceptions are thrown by checking if it contains the `errors` key. But it should be `error` as mentioned in the above response.